Solidify `invokeBoard` path.
