+++ packages/board-server/src/server/boards/invoke.ts
@@ -4,60 +4,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { getDataStore } from "@breadboard-ai/data-store";
-import { createLoader, inflateData } from "@google-labs/breadboard";
-import { run, type HarnessRunResult } from "@google-labs/breadboard/harness";
 import type { ApiHandler, BoardParseResult } from "../types.js";
-import {
-  BoardServerProvider,
-  loadFromStore,
-} from "./utils/board-server-provider.js";
-import { createKits } from "./utils/create-kits.js";
-import { formatRunError } from "./utils/format-run-error.js";
+import { loadFromStore } from "./utils/board-server-provider.js";
 import { verifyKey } from "./utils/verify-key.js";
 import { secretsKit } from "../proxy/secrets.js";
-
-export const invoke = async (
-  url: string,
-  path: string,
-  inputs: Record<string, any>
-) => {
-  const store = getDataStore();
-  if (!store) {
-    return;
-  }
-
-  const runner = run({
-    url,
-    kits: createKits([secretsKit]),
-    loader: createLoader([new BoardServerProvider(path, loadFromStore)]),
-    store,
-    inputs: { model: "gemini-1.5-flash-latest" },
-    interactiveSecrets: false,
-  });
-
-  for await (const result of runner) {
-    const { type, data, reply } = result as HarnessRunResult;
-    if (type === "input") {
-      await reply({ inputs });
-    } else if (type === "output") {
-      return inflateData(store, data.outputs);
-    } else if (type === "error") {
-      return {
-        $error: formatRunError(data.error),
-      };
-    } else if (type === "end") {
-      return {
-        $error: "Run completed without producing output.",
-      };
-    } else {
-      console.log("UNKNOWN RESULT", type, data);
-    }
-  }
-  return {
-    $error: "Run completed without signaling end or error.",
-  };
-};
+import { invokeBoard } from "./utils/invoke-board.js";
 
 const invokeHandler: ApiHandler = async (parsed, req, res, body) => {
   const { board, url } = parsed as BoardParseResult;
@@ -68,7 +19,13 @@ const invokeHandler: ApiHandler = async (parsed, req, res, body) => {
     res.end(JSON.stringify({ $error: keyVerificationResult.error }));
     return true;
   }
-  const result = await invoke(url, board, inputs);
+  const result = await invokeBoard({
+    url,
+    path: board,
+    inputs,
+    loader: loadFromStore,
+    kitOverrides: [secretsKit],
+  });
   res.writeHead(200, { "Content-Type": "application/json" });
   res.end(JSON.stringify(result));
   return true;
